+++ SSU/UMC/Spring_Boot_B/src/main/java/com/ssu/umc9th2/spring_boot_b/domain/notification/entity/Notification.java
@@ -1,6 +1,6 @@
 package com.ssu.umc9th2.spring_boot_b.domain.notification.entity;
 
-import com.ssu.umc9th2.common.base.BaseEntity;
+import com.ssu.umc9th2.spring_boot_b.common.base.BaseEntity;
 import com.ssu.umc9th2.spring_boot_b.domain.mission.entity.Mission;
 import com.ssu.umc9th2.spring_boot_b.domain.restaurant.entity.Restaurant;
 import com.ssu.umc9th2.spring_boot_b.domain.user.entity.User;
