+++ SSU/UMC/Spring_Boot_B/src/main/java/com/ssu/umc9th2/spring_boot_b/domain/review/entity/ReviewImage.java
@@ -1,6 +1,6 @@
 package com.ssu.umc9th2.spring_boot_b.domain.review.entity;
 
-import com.ssu.umc9th2.common.base.BaseEntity;
+import com.ssu.umc9th2.spring_boot_b.common.base.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
